pyproject.toml
[project]
name = "neural-memory-mcp"
version = "0.1.0"
description = "A neural memory MCP server simulating human-like memory for LLMs"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[project.scripts]
neural-memory = "neural_memory.__main__:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neural_memory"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
